Move `Tab` form handler to separate view
